Prevent pyconfig.h conflicts on multiarch systems
pyconfig.h is included in python-libs, as it is used on runtime.
The problem is pyconfig.h contains some platform-specific code and
it differs between platforms.

python-libs-*.x86_64  would conflict with python-libs-*.i686 and if they
are forced to coexist, wrong pyconfig.h may be used. This causes e.g.
rpm build to fail in a non obvious way.

The hack is to have architecture-specific pyconfig-*.h files and a ghost
symlink updated with python-devel install. I hope this works.